chore: hide survey
