+++ resources/js/stores/SurveyStore.js
@@ -1,0 +1,13 @@
+import { defineStore } from 'pinia';
+import { useStorage } from '@vueuse/core'
+
+export const useSurveyStore = defineStore('SurveyStore', {
+    state: () => ({
+        isDone: useStorage('isDone', false)
+    }),
+    actions: {
+        done() {
+            this.isDone = true
+        }
+    }
+});
